requirements.txt
flet
google-genai
python-dotenv
#